Add Image Path to Cars, update DrawPanel accordingly
The Link between cars and their images is now less fragile. DrawPanel
uses the car's position to update the image of that car accordingly. It
should now be possible to add more WorkShops and Cars (in the program /
main function). Known issues: CarController still only supports a single
WorkShop. Also the detection for a car running into a matching WorkShop
seems to have broken in the refactoring.